# Feat: Implement a PersistentDatastore by adding DiskUsage method