@lgierth 2018-02-16
> Slower Open() - it run Stat() on all files in the datastore

That's a bummer for the gateways -- can't have a full repo scan on every restart there.

Can we make the bookkeeping persistent so that it survives restart?