@hsanjuan 2018-02-16
@lgierth is it so bad? In my gateway tests a while ago it did take a few seconds, but it's once during boot and while ipfs is not hammering the disk at the same time.

That said, I can leave note of the size on a file, sure, and only re-scan when not present.