@Stebalien 2018-02-16
There are two downsides to caching disk usage:

1. Downgrade, do something, upgrade (as we haven't bumped the repo version).
2. It'll be even more out-of-date directory sizes (meh?).

An alternative is to calculate the initial size lazily. The first call to DiskUsage will be slow but it won't halt the world.