@hsanjuan 2018-02-16
Without looking at `Walk` internals, walking lazily probably suffers from race conditions where a file may be counted twice (one through Put as soon as we Open the datastore and the other one as the lazy walk finally gets to its folder). Also, the other way around with in the case of delete.

So it also has some downsides... (and the fact that it may return really low sizes at the beginning which may be to whoever is looking at that metric).

I lean towards caching even though it's not perfect...