@Stebalien 2018-02-16
> walking lazily probably suffers from race conditions where a file may be counted twice

Hm. You're right. We'd have to track which parts we've walked and which parts we haven't. That would be *horrible*.

Yeah, this sounds like the best solution. Note: If we want a more accurate solution, we could cache the size of the files and recompute the sizes of the directories but it's probably not worth it.