@hsanjuan 2018-02-16
> probably not worth it.

agreed. Will you be so kind to greenlight the code or leave feedback @Stebalien ?