@Stebalien 2018-02-16
So, reasoning through this carefully, there are a few annoying race conditions:

* If two users *write* the same block at the same time, we'll double-count the file.
* Same for two simultaneous deletes.
* A simultaneous write + delete could end up not recording the fact that we've freed disk space.

For the delete case, we should:

1. Compute the disk space.
2. Remove the file.
3. Update the usage iff *we* removed the file (the call to Remove succeeded).

For the write case... I think we'll have to use `os.Link(old, new)` and `os.Remove(old)`. That should work on all operating systems and should avoid overwriting the file. Note: this assumes that we never want to *replace* values in datastores. If we do, we'd have to do... we could do a remove and replace dance...

There are a few os-specific ways to make this faster but cross-platform low-level filesystem stuff always kind of sucks.