@hsanjuan 2018-02-19
@Stebalien you're right.

I have followed your indications. I'm not super happy though:

* Not allowing overwrite of values seems like a high price to pay and changes datastore behavior
* I cannot get my test of concurrent mixed put+delete over the same key to pass with `sync = false`. Sure, deviation is small, but it seems some sort of race exists still at a lower level. I'm happy ignoring this.
* I have tried to do the `remove and replace` dance. This introduces further race conditions (cannot remove and replace atomically) and I'm not sure it's possible to do that dance without making things horrible or recurring to `sync.Mutex`..

My feelings overall:

* DiskUsage bookeeping should not deeply impact datastore performance.
* DiskUsage should be more or less accurate for the common usage of this datastore, but it doesn't need to exact.

Thus, if we can live with the non-overwrite limitation i'd leave things like this, otherwise, I'd return to the buggy behaviour before (fixing the delete part, which is easy). What do you think?