@Stebalien 2018-02-19
> My feelings overall:

Overall, I agree. Performance is already poor enough as it is.

> Thus, if we can live with the non-overwrite limitation i'd leave things like this, otherwise, I'd return to the buggy behaviour before (fixing the delete part, which is easy). What do you think?

My worry is that simultaneous writes/deletes are probably quite common and, if they're not perfectly balanced, the estimated disk usage will skew over time. We might, actually, just want to use a synchronization mechanism. I wouldn't have a *global* lock, but we can track in-progress puts/deletes on a per-file basis. That is, I'd have a map (either a `sync.Map` or a map protected by a mutex) mapping file names to some form of synchronization object. Every operation would keep retrying itself until it either (a) observes itself fail or (b) observes some concurrent operation (or itself) succeed.

That is, given a timeline like:

```
<--op 1---> // takes the "lock"
  <--op 2-> // waits
   <-op 3-> // waits
```

If op1 succeeds, we can serialize these operations as `<op2><op3><op1>` (and say that `op2` and `op3` are valid for no observable time and therefor optimize them out :smile:). If op1 fails, we'll need to retry at least one of the other operations (in practice, they'll both retry and start at the top).

This should actually be *faster* than the existing implementation given concurrent operations and should be just as "correct".

Does this sound too complicated? Any better ideas?