@hsanjuan 2018-02-23
@Stebalien ok, I have implemented the approach you suggested. Works well for my tests (added an extra one too for putMany()). Hopefully we can settle here :)

The tricky code should be sufficiently commented, but let me know if something's not clear.

Note, Mac tests fail because `too many open files`. I'll tune that later.