@Kubuxu 2018-02-27
@hsanjuan SGWM, it also seems to me that triggering full re-scan on `ipfs repo fsck` would be good feature to have. This way if DiskUsage file diverged too far, fsck would fix it.

Any idea how to nicely wire it in?