@hsanjuan 2018-02-27
@Kubuxu that will need interface support in go-datastore. Should we have an `Fsck()` method for `PersistentDatastore` that does this (and potentially other things if needed)?