@magik6k 2018-02-27
The GC interface is wired into go-ipfs master already (merged in https://github.com/ipfs/go-ipfs/pull/4578)

I created 2 more interfaces in go-datastore when making the GC one - https://github.com/ipfs/go-datastore/blob/master/datastore.go#L88-L103. `Check` from `CheckedDatastore` is meant to be triggered on `ipfs repo fsck`. I'm not sure where `Scrub` should get invoked, probably separate command like `ipfs repo scrub`.

`ScrubbedDatastore.Scrub` is probably where we want to wire this into. (`Check` / `repo fsck` should be fast)