@whyrusleeping 2018-02-27
Could we get some benchmarks on how this will affect the overall performance of the datastore?