@hsanjuan 2018-03-01
@whyrusleeping @Stebalien my feeling here is that performance depends mostly on the OS and disk, and our changes (other than on a cold start), won't impact so much (and might improve the scenarios with concurrent writes to the same keys).

Let's have a look at existing benchmarks, which use random keys and non concurrent Puts, and sync = false.

### Using /dev/shm (in memory)

New:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          20000           1307640 ns/op
BenchmarkBatchedPut-8              20000           1358376 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    84.285s
```

Master:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          20000           1240002 ns/op
BenchmarkBatchedPut-8              20000           1721207 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    88.808s
```

Master again:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          20000           1838413 ns/op
BenchmarkBatchedPut-8              20000           1510638 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    96.605s
```

New again:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          20000           1083198 ns/op
BenchmarkBatchedPut-8              20000           1120730 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    75.054s
```

They don't seem to show anything terrible going on. If anything, things got faster **which I can't explain**. This happens consistently even after I repeat the benchmark several times.

---

### Testing with /tmp (spinning disk)

New:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8           3000            575120 ns/op
BenchmarkBatchedPut-8               5000           1446841 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    17.183s
```

Master:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          10000           2407771 ns/op
BenchmarkBatchedPut-8               5000           8187870 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    189.733s
```

Master again:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8          10000           3104441 ns/op
BenchmarkBatchedPut-8              10000           3186439 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    82.227s
```

New again:

```
goos: linux
goarch: amd64
pkg: github.com/ipfs/go-ds-flatfs
BenchmarkConsecutivePut-8           5000           2775705 ns/op
BenchmarkBatchedPut-8               5000           2180260 ns/op
PASS
ok      github.com/ipfs/go-ds-flatfs    32.512s

```

It would seem that the benchmarks try to adjust for more iterations in master (which means it was faster for lower N?). But iterating more eventually becomes a huge penalty so master is slower per loop.

---

From these benchmarks I can't see a negative impact of this branch over performance of `Put`. I think we are mostly subject to OS and Disk induced delays when interacting with the storage.